refactor: modernize codebase
