+++ .babelrc
@@ -1,0 +1,10 @@
+{
+  "presets": [
+    ["env", {
+      "targets": {
+          "node": "6.0"
+      }
+    }]
+  ],
+  "plugins": ["add-module-exports"]
+}
