+++ .eslintrc
@@ -1,9 +1,8 @@
 {
-  "extends": [
-    "prettier"
-  ],
+  "extends": ["prettier"],
   "parserOptions": {
-    "ecmaVersion": 2017
+    "ecmaVersion": 2017,
+    "sourceType": "module"
   },
   "env": {
     "es6": true,
