+++ .prettierrc
@@ -6,7 +6,7 @@
   "trailingComma": "all",
   "overrides": [
     {
-      "files": ".prettierrc",
+      "files": [".prettierrc"],
       "options": {
         "parser": "json"
       }
