+++ TODO.md
@@ -1,6 +1,4 @@
 1.  Document all functions
-1.  Transpile with babel
 1.  Implement semantic-version
 1.  Implement code coverage reports
 1.  Support unsupported releases
-1.  pass download options to axios
