--- bin/lnd-install
@@ -1,9 +1,0 @@
-#!/usr/bin/env node
-
-/**
- * Script that will install lnd.
- */
-
-var path = require('path')
-
-require(path.join(__dirname, '..', 'scripts', 'install'))
