+++ src/bin/lnd
@@ -5,11 +5,11 @@
  * forwarded to and from the child process.
  */
 
-var path = require('path')
-var spawn = require('child_process').spawn
-var lnd = require(path.join(__dirname, '..', 'lib', 'extensions'))
-var binPath = path.join(__dirname, '..', 'vendor', 'lnd' + lnd.getBinaryFileExtension())
+import path from 'path'
+import { spawn } from 'child_process'
+import lnd from '../lib/extensions'
 
+var binPath = path.join(__dirname, '..', '..', 'vendor', 'lnd' + lnd.getBinaryFileExtension())
 var args = process.argv.slice(2)
 
 // For Node 0.6 compatibility, pipe the streams manually, instead of using
