+++ src/bin/lnd-install
@@ -1,0 +1,11 @@
+#!/usr/bin/env node
+
+/**
+ * Script that will install lnd.
+ */
+
+import path from 'path'
+
+import { install } from '../lib/install'
+
+install()
