+++ src/lib/manifest.js
