+++ src/lib/util/proxy.js
@@ -13,6 +13,6 @@
  * @return {String} the proxy configured by npm or an empty string
  * @api private
  */
-module.exports = function() {
+export const proxy = function() {
   return process.env.npm_config_https_proxy || process.env.npm_config_proxy || process.env.npm_config_http_proxy || ''
 }
